Renamed draw to editor
