+++ src/components/editor/index.js
